backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
pandas
django-extensions
# For testing
coverage

# for postgres and django
psycopg2

# ImageField
Pillow

# For caching
django-redis


# For gtfs
gtfs-realtime-bindings

